Make hardhat deploy test an integration tests
This will run the test in isolation and prevent it from failing.

Also combines the other integration tests in a single binary. This
should reduce compilation and test time. See the "Modularizing tests"
section of https://mudit.blog/rust-tips-and-tricks/.
