+++ contracts/rust/tests/combined/smoke_tests_deployed_contract.rs
