+++ contracts/rust/tests/combined/wrapped_assets_can_be_frozen.rs
