+++ contracts/rust/tests/combined/wrapping.rs
